requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ckeditor==6.3.2
django-js-asset==2.0.0
Pillow==9.5.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
